django_backend/project/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
choicesenum==0.7.0
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.11
django-allauth==0.47.0
django-cors-headers==3.11.0
django-exiffield==3.0.0
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
jsonfield==3.1.0
Markdown==3.3.6
oauthlib==3.1.1
Pillow==9.0.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0